Enable OMH support in Android Telephony
Support the Open Mobile Handset initative in android telephony. As per
the OMH specification in 3GPP2, Data profiles defined for specific
services like LBS, TETHERED and UNSPECIFIED are defined in the
RUIM/CSIM cards. These profiles are read from the android telephony
and used for creating the data calls when the right radio technology
(CDMA/DO) is used and when the feature is enabled.

* Supports multiple profiles for a particular OMH service type

* Supports multiple OMH profiles with same priority

* In case of OMH, for all profiles supporting SUPL, arbitration
should choose the lowest priority profile when making data calls

* This feature supports the co-existance of other radio technologies in
an OMH enabled device. Even with the OMH feature enabled, the device
can operate with default profiles when OMH profiles are not (yet)
present.

* Use PROFILE_TYPE_CDMA if no suitable PROFILE_TYPE_OMH available

* Fix phone crash:
Default code triggers setupDataCall even for cdma detachet event.
In a handoff scenario cdma detach event is sent before the objects
are initialized (profiles being read) and CdmaDCT attempts a call
with null DataProfile object.

* Add null checks in CdmaDataConnection

Change-Id: I7f58c9a4ba434b1fde01fa4a0abfaf0ce4cacbea