Telephony: Add 'ServiceCategory' to SMS finger print
In case of certain broadcast messages sent back to back, say if,
service category 0x0002 Administrative and service category
0x0003 Maintenance are sent in sequence,
one after the other, service category 0x0003 fails to be displayed
on the UI although modem sends it to telephony
layer. This is because telephony rejects/discards the new
incoming message - 0x0003 as its finger-print matches exactly
with the previously sent message 0x0002.

Change-Id: Id48169924605f423ee97a0866e754462f8c4c30d