Fix bug. Add to Delete uncompleted concatenated message.
We consider the way to delete uncompleted concatenated Message.
There is 2 ways. Implement the following case in com.android.internal.telephony.SMSDispatcher.processMessagePart()
1. Newly received into the existing data space -Compare the stored data and the new data.
a. if matched , the new data is the duplicated data, so it should be discarded.
b. if not matched, the stored data is obsoleted, so the ALL stored data related the concatenation reference number should be deleted.
2. Newly received into the lacked space -Compare timestamp of the stored data and the one of the new data
( The timestamp is incremented by 1 according to the TS23.040. So, the stored segments are obsoleted if delta>=256. The delta same "VALID_TIMER_CONCAT_SEGMENT" variable in source code)
a. if the delta <256 then the New data is the valid segment, so it should be concatenated.
b. if the delta >=256 then, the stored data is obsoleted, so the ALL stored data related the concatenation reference number should be deleted.
To discard the ALL sored data related the concatenation reference number, we use "discard" variable.
Signed-off-by: kyunga kim <kyunga1.kim@lge.com>

Change-Id: I8a69503af90cf6d52f7646086b1cafca07a1ddf2