Delete incomplete concatenated message.
There are two cases to handle.
1. Compare the stored data and the new data.
And then the stored data is obsoleted., if not matched.
2. Compare timestamp of the stored data. and the new data.
And then the stored segments are obsoleted if delta>=256.
The timestamp is incremented by 1 according to the TS23.040.
("delta" same "MAX_CONCATENATED_SEGMENT_TIME" variable in source code)
According above 2 cases,
The ALL stored data related the reference number should be deleted.
Signed-off-by: kyunga kim <kyunga1.kim@lge.com>
Change-Id: I255bd3b77d4600fe8c39d4fea09ecd8f39e1b2cf