Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

(1 inline comment)

This change adds support for the method of encoding email headers defined in 3GPP 23.040 section 3.8. It does not add support for the completely different mechanism described in 9.2.3.24.11.

Supporting the email header encoding in 9.2.3.24.11 will require adding support to SmsHeader to parse IE's of type ELT_ID_RFC_822_EMAIL_HEADER, which contains a single octet describing the number of characters (septets, octets, or UCS2 characters, depending on user data encoding) in the user data for the RFC 822 header. Then parseMessageBody() will have to parse the first N characters as regular RFC 822 headers when this IE is present in the UDH, and the remainder of the message will be the email body.

I think this change is worthwhile to support the email gateway encoding defined in section 3.8, but the commit message and code comments should be changed to reflect that this is not the same as the RFC 822 header encoding described in 9.2.3.24.11, which is still not supported with this change. I think the original JavaDoc comment for SmsMessageBase.extractEmailAddressFromMessageBody() could be the cause for the confusion, since the last line mentions section 9.2.3.24.11 while the comment is describing section 3.8.

Patch-set: 2
Reviewer: Jake Hamby <1003544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Verified=0