Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

The commit message should clarify that this is a fix for dual-mode CDMA/LTE devices (and there should be a comment in the code to explain the change as well). On HSPA/LTE devices, CellLocation should always be an instance of GsmCellLocation, unless there's another change to create a separate CellLocation subclass for LTE that I haven't seen yet.

Patch-set: 1
Label: Code-Review=-1