Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I don't understand the reason for this patch: CMAS messages should be handled in CdmaSMSDispatcher.dispatchMessage() by the call to sms.parseBroadcastSms() and dispatchBroadcastMessage(message), and sms.getIncomingSmsFingerprint() should only be called when the message is not a broadcast SMS.

In fact, I recently changed cdma.SmsMessage.getMessagetype() to return MESSAGE_TYPE_BROADCAST whenever the serviceCategory != 0, to work around a RIL bug that was not setting the messageType correctly for incoming CB and SMS, so the only time that getIncomingSmsFingerprint() will be called in the current code is when the service category is 0.

So I don't understand under what circumstances the service category would be non-zero where the message would not be handled by the broadcast SMS code which does not call getIncomingSmsFingerprint(). BTW, the CellBroadcastReceiver app handles detection of duplicate cell broadcasts, and they aren't detected in the framework.

Patch-set: 1
Label: Code-Review=-1