Telephony: Mark SIM SMS as read after importing it
The GCF test case 8.2.2 checks if the status byte of the SMS is marked as read
after reading the SMS from SIM. The status byte is marked as read after the
SMS is imported from the SIM if it is not already marked.
Fix to check for null IccFilehandler which can happen if card is not present.

Extract common functions from Sim/RuimSmsInterfaceManager to their parent
class IccSmsInterfaceManager.

Change-Id: I4f226bc873786904527d7d822a67dfc14f05a327