Update patch set 1
Patch Set 1:

Hi Jake,

You are right... This fix doesn't apply to JB.  Question regarding your comment:  "BTW, the CellBroadcastReceiver app handles detection of duplicate cell broadcasts, and they aren't detected in the framework."  Tip of aosp does not have this change you are referring to.  But we did find it was part of android-4.1.1_r2 release and on.  Just want to verify that "Fix CellBroadcastReceiver to ignore duplicate broadcasts." is not reverted.  It is still valid change, right?  Thanks, Rika

Patch-set: 1
Label: Code-Review=0