Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

What bug are we addressing here? I want to be careful that we minimize assumptions about the correlation between radio being on/off/low-power and the SIM being on/off/present/absent. And I'm worried this is adding to the correlation.

Patch-set: 1
Label: Code-Review=-1