Telephony: Update the state of FDN Service correctly
FDN Service state was derived based on the state of PIN2. Update the
state of FDN service based on the FACILTY_LOCK messages instead.

Telephony: Change the file path to FDN,MSISDN for USIM and CSIM
Correcting the file path to EF_FDN and EF_MSISDN from Telecom DF to ADF
for USIM and CSIM.

Telephony: Correcting the file path to EF_ADN and EF_MSISDN
This fix corrects the file path to EF_ADN and EF_MSISDN.

Change-Id: I099698d0031f980c3883c17f3c170fbcadd7fb84