Telephony: Set the operator numeric property correctly.
In RuimRecords, changes to set the PROPERTY_ICC_OPERATOR_NUMERIC
only if the operator is not null.  This is to ensure that the
property is set from ro.cdma.home.operator.numeric even if the
GET_IMSI failed in case of CDMA RUIM mode.

In SimRecords, changes to make sure that the PROPERTY_ICC_OPERATOR_NUMERIC
is not reset, to allow the property to be updated from
the ro.cdma.home.operator.numeric in case of CDMA RUIM mode with
more than one applications in the card.

Change-Id: I8fc81b0224fcb3205e08c0cad2988158ee5c0d9a