Telephony: notify the link properties changed in CdmaDCT
The CdmaDCT is not notifying the data connection state changed
when the UNSOL_DATA_CALL_LIST_CHANGED with the same active data
connection but the link properties are changed. Changes are made
to notify the updated link properties from the CdmaDCT on
processing the UNSOL_DATA_CALL_LIST_CHANGED notification.

Change-Id: I5522407600294e3e35705af6775493024b26bb4e