Telephony: Enabling Network Personalization PUK in Framework
Network personalisation is the feature were a SIM is restricted
to be used in a specific operator networks only. If an unsupported
SIM is used in the device, the user is asked to enter an unlock
code, so that the SIM can be used in the network. If the user enters
wrong PIN for a preprogrammed number of times, the device is PUK
locked and needs to be taken to the customer care. Android by default
supports entering the PIN, but doesn't inform the user if the device
is PUK locked.

This patch handles the NETWORK PUK required status from modem in
command exception, handles the new event registration and notifications
in IccCardProxy and UiccCardApplication.

Note: This patch is related to a change in platforms/frameworks/base,
		platform/frameworks/opt/telephony and platform/packages/apps/Phone

Change-Id: Id3b93f9085e7508770eb232ba50b728ffa4d54e2
Author: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 18871