Telephony: Add emergency call status intent
There are system services which needs to know the emergency
call status before taking any drastic actions.

With this patch, android.intent.action.EMERGENCY_CALL_STATUS
will be broadcasted when an emergency call is initiated and
terminated.

Note: This patch is related to a change in platform/packages/apps/Phone
		and platform/frameworks/opt/telephony

Change-Id: I2f3d27831cb44cddc2241f1c29a61440d3caa194
Author: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Shuo Gao <shuo.gao@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 55480