Update patch set 2
Patch Set 2:

(7 comments)

Sorry Ying Wei,  I didn't look at this for a while.  I will try to be more responsive while we work it out.

I reviewed it this morning and it appears to me know that mRequestMessagePending should simply be protected by the mWakeLock sync blocks.  It is only used to determine when the wake lock should be released and so this association with the sync block makes sense.

I think we've got a similar problem with mRequestMessagesWaiting.  It should also only be altered inside sync(mWakeLock) (clearRequestsList for example clears it from a separate thread).

Let me know what you think.

Patch-set: 2