Update patch set 3
Patch Set 3:

So the fix seems reasonable, we didn't have AdnRecord.aidl in the source file list. As Ying said in the google groups discussion:

https://groups.google.com/forum/#!topic/android-building/4ewpJy4eQ0g/discussion

Its probably because we should have added a cleanspec step. 

Its also peculiar that this has not caused a problem before because AdnRecord.aidl has been missing from the beginning. Either some other build magic causes the AdnRecord.aidl to be processed or its just not needed.

Patch-set: 3