Update patch set 2
Patch Set 2:

Yes, I mean have one lock on mWakeLock for access to mWakeLock, mRequestMessagesWaiting and mRequestMessagesPending.

Yes, that does mean it will serialize the threads.  That is the point.

Currently you can have ...Pending checked&set simultaneously on your message handler and any thread calling the public Send functions.  The ...Pending variable is used to decide if we can release the wake lock, which is already protected by sync(mWakeLock) and so it seems reasonable to extend that protection the the check&set blocks of ...Pending.

Same applies to ...Waiting.

Adding separate locks to these variables will complicate the code and potentially lead to deadlock.

Patch-set: 2