Update patch set 2
Patch Set 2: Verified-1

Imagine two threads calling acquireWakeLock and releaseWakeLock at about the same time.

The Acquire thread hits a little first - grabs the sync on mWakeLock, acquires the wake lock and then releases the sync on mWakeLock.

Then we switch

The Release thread grabs the sync on mPendingLock, finds mRequestMessagesPending is 0, sets hasMessagePending to false, releasing mPendingLock, grabs the sync on mWakeLock and releases the wakelock.

We then switch back

mPendingLock's sync is grabbed by the aquire thread and mRequestMessagesPending is inc'd

Now because of a sync bug we don't have wakelock held when it should be.  Had the check+act been synched by a single lock in releaseWakeLockIfDone this couldn't have happened.

If you want, I'll post my fix in a separate change so you can look at it.

Patch-set: 2
Label: Verified=-1