Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

Please change the other length readByte() to readUnsingedByte() on line 506.

Also, any readByte() which is assigning to an int should be changed to readUnsignedByte. If you don't wish to do that we'll do it internally. In additinon there are a few casts of (byte) which are assigning to integers, these casts should be removed. Specifically lines 152, 153 and 155.

Thanks for finding this bug!

Patch-set: 3
Label: Code-Review=-1