Update patch set 1
Patch Set 1:

I uploaded the unit tests that I created to verify this patch.

https://android-review.googlesource.com/#/c/49461/

The new unit tests require support for Shift-JIS encoding as well as decoding, which can be added to this patch in the same way that Shift-JIS decoding was added, as I mentioned in a previous comment.

Japanese characters require 2 bytes in both Shift-JIS and UTF-16, so the only benefit of Shift-JIS is that ASCII and half-width katakana characters (plus the yen symbol) are encoded in 1 byte.

In a future release, we could enable Shift-JIS encoding for messages that would benefit from it, but this will have the same problem as the Turkish (and other national language) 7-bit encodings for GSM SMS, which is that the receiving phone must support the encoding. The Turkish 7-bit tables are only enabled for encoding on Turkish SIM cards by default (or they can be enabled for a specific device build using a config property), to comply with Turkish law requiring their use. All devices can decode messages encoded with any of the national language tables, but they are only used for encoding when specifically enabled by a custom build property or the detection of a Turkish SIM card.

So the reasons that I think this patch should add an else branch to BearerData.encodeUserDataPayload() to support encoding with Shift-JIS when a UserData object is passed which specifically requests it are for symmetry and because it makes it easier to write test cases such as the ones I just uploaded. However, we shouldn't make any other code changes to enable the use of Shift-JIS for the normal code paths.

In the future, if a Japanese CDMA carrier specifically requests Shift-JIS to be enabled for encoding, we can add code to enable it under specific conditions (e.g. registered network has Japanese MCC or device has Shift-JIS encoding enabled via build property), as we did for the Turkish case.

Patch-set: 1