Update patch set 2
Patch Set 2:

OK, Patch set #1:

Acquire is called first - the sync of mWakeLock is acquired, the wake lock is acquired and the sync of mWakeLock is released. 

We then switch threads to a caller of releaseWakeLockIfDone.  The sync of mWakeLock is aqcuired.  Pending is checked and is 0.  The WakeLock is released.  The sync of mWakeLock is released.

We switch back and Pending is incremented.

We now think the WakeLock is held (it should be) but it is not.

Patch-set: 2