Fix for Call forwarding registration using MMI codes
We don't register for call forwarding when *SC*SI# is
dialed where SI is the number to which call should be
forwarded. We activate instead a previously registered
number.

As per 3GPP specification 22.030 sec. 6.5.2
The UE shall determine from the context whether, an entry
of a single *, activation or registration was intended.
For example, a call forwarding request with a single *
would be interpreted as registration if containing a
forwarded-to number, or an activation if not.

b/7903480

Change-Id: I15e12efd5f1691fae166f2e5bba77323766e5801