Update patch set 1
Patch Set 1:

If preferred network type is changed in the test menu, the selected network type value is passed to Qualcomm modem layer.
And modem saves the preferred network type and starts to search network according to the passed network type.

In this case, modem will try to search WCDMA network only and not search 2G network because the preferred network type is changed to "WCDMA only".
So, modem will not attach 2G network even if framemwork sends to setup data call to modem when you leave wifi area.
This is normal behaivor.

Patch-set: 1