Wrong terminal response for proactive command length errors
This fix is at least reported by type approval, can not
guarantee that it exists "in the wild" however the
specification is clear.

If the total lengths of the SIMPLE-TLV data objects are not
consistent with the length given in the BER-TLV data object,
then the whole BER-TLV data object shall be rejected. The
result field in the TERMINAL RESPONSE shall have the error
condition "Command data not understood by ME".

Change-Id: I128442ea4755679b0e5524a31c01b3c690e45ab0