Update patch set 5
Patch Set 5: Code-Review-1

(3 comments)

There's some extra/missing whitespace in the code formatting (see inline comments), but I also have a concern about the implementation.

It took me a little while to figure out where the initial TP-MR is assigned. The framework sets the TP-MR byte of the PDU to 0 in .gsm.SmsMessage.getSubmitPduHead(), so the RIL is responsible for filling it in. Then we get the messageRef from the response to RIL_REQUEST_SEND_SMS and this patch sets the TP-MR byte to the same value before resending.

So my only concern is that the RIL must be careful not to overwrite TP-MR with a different value when this code path is executed. Because 0-255 are all valid values for TP-MR, the correct RIL behavior is not obvious when TP-MR is zero. I suspect that at least some vendor RILs may be assigning a value to TP-MR regardless of the value passed in the PDU. 

The correct RIL behavior would probably be to assign a new TP-MR value when TP-RD is 0, and to use the value from the PDU when TP-RD is 1. Rika, is that your understanding as well, or is there something I'm missing?

The change isn't merging cleanly into our internal tree because of some variable name changes that Wink made for consistency with Android naming conventions. It's easy enough to cherry-pick into our internal tree by updating the variable names and changing Log to Rlog, but we'll have to do that by hand.

Patch-set: 5
Reviewer: Jake Hamby <1003544@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1