Update patch set 5
Patch Set 5:

Hi Jake, In our solution, it is actually not the RIL, but rather the WMS module in modem that looks up next TP-MR to be used from SIM @power up and supplies when TP-RD is 0.  And RIL is just a pass through.  When MO SMS failes, failed TP-MR value is propagated from WMS -> RIL -> Telephony.

What you said about "I suspect that at least some vendor RILs may be assigning a value to TP-MR regardless of the value passed in the PDU." may be true, and those vendors must fix their RIL if they want this change to take effect in setting TP-RD and TP-MR correctly.

We'll fix the white spacing.

So, are you satisfied with my response or do you have any other concern?

Thanks,
Rika

Patch-set: 5