Enhance the cdma 3-way call
With sprint network, CDMA 3-way call need a blank flash to hold 1st call,
and the 2nd call should be send to network after a bogus timer. And this
has no harm to other networks which do not require the extra blank flash,
so this could be a common process for all CDMA 3-way call.

Change-Id: I4672b25aabfffb3e1ae7593ba3b0352be80f7463

Conflicts:

	src/java/com/android/internal/telephony/cdma/CdmaCallTracker.java