Invalid call forwarding indicator status check
According to ETSI specification, it is needed to check the value of
the first byte of EF_CFIS MSP, which is between 1 and 4. However,
existing many SIM cards don't follow ETSI specification, and hence,
CFU (Call Forwarding Unconditional) function is usually failed.

This patch deletes the check for the first byte of MSP so that
existing SIM cards can use CFU function.

Change-Id: I78fe0f3ddc4d3650e8cae0efea80e0ed84917c2a