Update patch set 1
Patch Set 1: Code-Review-1

This looks like it will be very useful. I'll add some more comments after I've had time to test and review the code more thoroughly.

A few quick questions:

1) Does this code require any changes to the telephony classes themselves? I didn't see any other recent uploads, and I'm a little bit surprised if you were able to make this work without any changes to the classes that you're testing, because the last time I looked into trying to do something like this, there were some singletons and static methods that made it difficult to test the pieces with mock CommandsInterface and other classes like you're doing.

2) Are there any special commands or instructions to run these tests? It looks like I should be able to do "runtest frameworks-telephony" but perhaps the specific classes need to be specified with the -c flag?

The commit message will need a little bit more detail. I think we need at least a few sentences explaining the rationale for the new framework (perhaps a summary of the proposal that Ayako-san uploaded, without the diagrams) and a short list of the most important new classes and their roles.

I'd like to see if we can create a set of test cases that could run in this environment and also in a mode that would use the actual RIL.java plus a signaling tester controlled by a replacement class for RILSimulator.java.

I have some test code that I've been working on (to use Anritsu MD8475A at first, but designed to be able to add support for other test boxes) that I will upload once it's ready, hopefully some time next week.

I've been thinking about whether we can build a framework that would allow us to write test cases that can run in both a mock environment like this and also in a call box test environment. The majority of code in this CL is probably only useful for frameworks-telephony tests because the purpose of CTS testing is not to require a particular internal implementation, but to ensure the best user experience and application compatibility, so CTS tests are "black box" vs. these "white box" test cases. We definitely need more of both types of tests in telephony.

Although the implementation details are different, it seems to me that with the right level of abstraction, there are many tests that could be written in such a way that they could function in both types of test environments. With either the mock classes or while running on a test network, at a high level the test is specifying what state changes to expect (and what state changes we should not expect), then triggering either Android or the remote side (the mock classes or the signaling tester) to perform various events and then watching the events and state changes that we see on the opposite side.

Patch-set: 1
Label: Code-Review=-1