Update patch set 1
Patch Set 1:

This is part of a two project commit so the duplication is to set the same Change-Id.

There is an ongoing change proposal being evaluated that will have significant differences from the current one. Should it be closed and create a new Change-Id or reopen this one and abandon https://android-review.googlesource.com/57916/ to further ammend the new changes?

Patch-set: 1
Reviewer: Antonio Marin <1021495@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0