Update patch set 1
Patch Set 1:

Hi Björn,

I understand your concerns. My understanding of the situation is that, while a sophisticated end-user such as yourself may be annoyed at certain carriers inflating HSPA+ to "4G" status (and LTE to "4G LTE"), for example, that battle has already been lost. The average end-user now expects to see whatever label their carrier is advertising, whether it be "4G" or "H+" or "4G LTE", and may return the phone as "defective" if it doesn't show the same label as other phones on that carrier show for that network type.

At this point, whatever labels that each carrier wants to assign to their various radio technologies, the number of signal bars they want to use, and other details, has been standardized across all devices, and all Android OEMs as well as Apple, Microsoft, etc. have had to follow suit or else they won't be able to sell devices to that carrier.

There are several other changes included here that will be of benefit to end-users, by making the status icons more readable, or are required by laws in various regions, so we will try to include the changes that are good for the user into the Nexus and AOSP builds, while minimizing the operator-specific inconsistencies that you're concerned about.

The primary rationale for contributing this code to AOSP to enable all of the operator-specific customizations is that these are changes that all major Android OEMs are now having to make in their private trees, and then have to merge into each new AOSP release, which can delay updating their phones to new versions of the OS by several weeks.

So even though we might not be happy about operators demanding these various tweaks, I hope we can agree that if OEMs are forced to make them anyway in order to sell carrier-branded devices, we can at least try to get a single framework into AOSP to make their lives easier, so they can get new versions of Android into users' hands with minimal delay.

I hope that explanation is helpful, and I will review and comment on the patch set later this week. Thanks!

Patch-set: 1