UICC : add support empty alpha id for BIP
"If the alpha identifier is provided by the UICC and is a null data
object (i.e. length = '00' and no value part), this is an indication
that the terminal should not give any information to the user
or ask for user confirmation;
if the alpha identifier is not provided by the UICC, the terminal may give information to the user;
if the user does not reject the channel, the terminal shall then set up a channel;"
"
- 6.4.27.2 OPEN CHANNEL related to packet data service bearer, ETSI TS 102.223

The device should set up a channel when the command is OPEN CHANNEL which has no alpha tag or null alpha tag data.

Bug: 9671401
Change-Id: Iece01f3775baf2c6438776f899ed53b1f1fb942f