Update patch set 1
Patch Set 1: Code-Review+1

Per 3GPP specification 102.223, "if the alpha identifier is not provided by the UICC,  the terminal MAY give information to the user". Do we need to handle that use case as well?

Suggest, we introduce a parameter in config.xml to make this behavior configurable.

Patch-set: 1
Label: Code-Review=+1