UICC : fix read EF Image Instance
The EFs(4Fxx) path under DF Graphics are not distinguish
with the EFs(4Fxx) path under DF Phonebook.
The getEFPath() is not able to return correct path.
So, the path should be defined as MF_SIM + DF_TELECOM + DF_GRAPHICS.

And some duplicated codes are removed on handleMessage() of IccFileHandler class.

Bug : 9767739

Change-Id: I754ec5c40cff99adef11a9dfdc31b8e0446befbb