UICC : fix read EF Image Instance
The EFs(4Fxx) path under DF Graphics are not distinguish
with the EFs(4Fxx) path under DF Phonebook.
The getEFPath() is not able to return correct path.
So, the path should be defined as MF_SIM + DF_TELECOM + DF_GRAPHICS.

EF_IMG is a linear fixed EF. So, the result of loading EF_IMG should be
processed as a LoadLinearFixedContext. So, it is needed to calculate the
number of EF_IMG records. If those changes are added, the changes are
duplicated with the codes of EVENT_GET_RECORD_SIZE_DONE. So, the codes
of EVENT_GET_RECORD_SIZE_IMG_DONE are removed and the event is treated
by the logic of the EVENT_GET_RECORD_SIZE_DONE. And then remove incorrect handler
events(EVENT_READ_IMG_DONE and EVENT_READ_ICON_DONE) are moved to the handler
events which have the procedure for loading same type EFs (EVENT_READ_RECORD_DONE
and the EVENT_READ_BINARY_DONE).

Bug : 9767739

Change-Id: I754ec5c40cff99adef11a9dfdc31b8e0446befbb