Update patch set 7
Patch Set 7:

Sim Contacts are examples for ADN.

UI not showing the record (contact)from SIM when entered record number# in dialer app. In case of GSM it is working fine, only in case of CDMA it is not working. 

Yes, this is not allowed on CDMA till now.

Patch-set: 7
Reviewer: Rakesh Pallerla <1021368@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0