Stk Application is not seen after sim Hot Swap
Upon sim removal, CatService is disposed and set
to null in UiccCard. On dispose, CatService
unregisters from the ril events. When the sim is
reinserted, existing CatService is reinitialized but
it doesn't register to ril events again. Due to this,
proactive commands are not notified to CatService.

If sRegisterForRilEvents is set, then register to ril
events on CatService reinitialization.

Change-Id: Ibd6c6d794ed53ce0df95e6b3226dbf9f96062878
Signed-off-by: Jeevaka Badrappan <jeevaka.badrappan@intel.com>
Signed-off-by: Arun Ravindran <arun.ravindran@intel.com>
Signed-off-by: Bruce Beare <bruce.j.beare@intel.com>
Signed-off-by: Jack Ren <jack.ren@intel.com>
Author-tracking-BZ: 127194